requirements.txt
streamlit
python-dotenv
pdfplumber
tqdm
cohere
pinecone
groq
numpy
PyMuPDF
